added a new mean embed version and debugging files

+++ examples/mean_embed/include/coll_avoid_main.h
@@ -2,7 +2,6 @@
 #define __COLL_AVOID_MAIN_H__
 
 #include <stdint.h>
-
-void getThrusts(int32_t*);
+void getThrusts(int16_t*);
 uint8_t isHighLevel();
 #endif
